Move FooterLinks to their own file
