+++ src/components/Footer/links.js
@@ -10,57 +10,57 @@ import linkedin from '../../images/linkedin-icon.svg'
 const social = [
   {
     label: 'Twitter',
-    link: 'http://twitter.com/yldio',
+    to: 'http://twitter.com/yldio',
     img: twitter
   },
   {
     label: 'Github',
-    link: 'http://github.com/yldio',
+    to: 'http://github.com/yldio',
     img: github
   },
   {
     label: 'Youtube',
-    link: 'https://www.youtube.com/channel/UCjCCJWM2iVVhqjKzJ-Y9MvA',
+    to: 'https://www.youtube.com/channel/UCjCCJWM2iVVhqjKzJ-Y9MvA',
     img: youtube
   },
   {
     label: 'Medium',
-    link: 'https://medium.com/yld-engineering-blog',
+    to: 'https://medium.com/yld-engineering-blog',
     img: medium
   },
   {
     label: 'Instagram',
-    link: 'https://www.instagram.com/yld.io/',
+    to: 'https://www.instagram.com/yld.io/',
     img: instagram
   },
   {
     label: 'Linkedin',
-    link: 'https://www.linkedin.com/company/yld/',
+    to: 'https://www.linkedin.com/company/yld/',
     img: linkedin
   },
   {
     label: 'Dribbble',
-    link: 'https://dribbble.com/yld',
+    to: 'https://dribbble.com/yld',
     img: dribbble
   },
   {
     label: 'Behance',
-    link: 'https://www.behance.net/YLDio',
+    to: 'https://www.behance.net/YLDio',
     img: behance
   }
 ]
 
 const gdpr = [
   {
-    text: 'Cookies',
+    label: 'Cookies',
     to: '/cookie-policy'
   },
   {
-    text: 'Privacy policy',
+    label: 'Privacy policy',
     to: '/privacy-policy'
   },
   {
-    text: 'Data-retention policy',
+    label: 'Data-retention policy',
     to: '/data-retention-policy'
   }
 ]
